# Remove N+1 query loading tags for each order cycle exchange