--- a/app/controllers/admin/subscriptions_controller.rb
+++ b/app/controllers/admin/subscriptions_controller.rb
@@ -162,25 +162,20 @@ module Admin
       [:index]
     end
 
+    def managed_enterprise_id
+      Enterprise.managed_by(spree_current_user).select('enterprises.id').
+        find_by(id: params[:enterprise_id])
+    end
+
     def subscription_params
       @subscription_params ||= PermittedAttributes::Subscription.new(params).call.
         to_h.with_indifferent_access
     end
 
     def payment_method_tags_by_id
-      payment_method_tags = ::ActsAsTaggableOn::Tag.
-        joins(:taggings).
-        includes(:taggings).
-        where(taggings: { taggable_type: "Spree::PaymentMethod",
-                          taggable_id: Spree::PaymentMethod.from(Enterprise.managed_by(spree_current_user).
-                          select('enterprises.id').find_by(id: params[:enterprise_id])),
-                          context: 'tags' })
-
-      payment_method_tags.each_with_object({}) do |tag, hash|
-        payment_method_id = tag.taggings.first.taggable_id
-        hash[payment_method_id] ||= []
-        hash[payment_method_id] << tag.name
-      end
+      @payment_method_tags_by_id ||= BatchTaggableTagsQuery.call(
+        Spree::PaymentMethod.from(managed_enterprise_id)
+      )
     end
   end
 end