--- a/app/queries/batch_taggable_tags_query.rb
+++ b/app/queries/batch_taggable_tags_query.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+class BatchTaggableTagsQuery
+  def self.call(taggables)
+    tags = ::ActsAsTaggableOn::Tag.
+      joins(:taggings).
+      includes(:taggings).
+      where(taggings:
+        {
+          taggable_type: taggables.model.to_s,
+          taggable_id: taggables,
+          context: 'tags'
+        })
+
+    tags.each_with_object({}) do |tag, indexed_hash|
+      tag.taggings.each do |tagging|
+        indexed_hash[tagging.taggable_id] ||= []
+        indexed_hash[tagging.taggable_id] << tag.name
+      end
+    end
+  end
+end