--- a/app/serializers/api/admin/order_cycle_serializer.rb
+++ b/app/serializers/api/admin/order_cycle_serializer.rb
@@ -34,7 +34,8 @@ module Api
 
         ActiveModel::ArraySerializer.
           new(scoped_exchanges, each_serializer: Api::Admin::ExchangeSerializer,
-                                current_user: options[:current_user])
+                                current_user: options[:current_user],
+                                preloaded_tags: BatchTaggableTagsQuery.call(scoped_exchanges))
       end
 
       def editable_variants_for_incoming_exchanges