--- a/spec/queries/batch_taggable_tags_query_spec.rb
+++ b/spec/queries/batch_taggable_tags_query_spec.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe BatchTaggableTagsQuery do
+  it "fetches tags for multiple models in one query" do
+    customer_i = create(:customer, tag_list: "member,volunteer")
+    customer_ii = create(:customer, tag_list: "member")
+    customer_iii = create(:customer, tag_list: nil)
+
+    tags = BatchTaggableTagsQuery.call(
+      Customer.where(id: [customer_i, customer_ii, customer_iii])
+    )
+    expect(tags).to eq(
+      {
+        customer_i.id => ["volunteer", "member"],
+        customer_ii.id => ["member"],
+      }
+    )
+  end
+end