#### What? Why?

- For https://github.com/openfoodfoundation/openfoodnetwork/issues/5213 

When an order cycle is loaded in the admin area a separate query is made to fetch the tags for each and every exchange that belongs to the order cycle. This changes things so tags are fetched for every order cycle exchange in a single query.

#### What should we test?

1. Find an order cycle which has multiple exchanges with tags, the more the easier it might be notice a difference, there should be less queries in any case.
2. Compare the before and after performance of `/admin/order_cycles/:id.json`. This URL is loaded via AJAX on the edit order cycle page.

#### Release notes

Remove N+1 query loading tags for each order cycle exchange

Changelog Category: Technical changes