@cillian 2022-11-18
Thanks for the review @jibees I think `spec/support/query_counter.rb` might be ok in this case because it's more of a helper class like `spec/support/test_summarizer.rb`