@jibees 2022-11-18
You're absolutely right @cillian . Moving to Test Ready then!