@drummer83 2022-11-23
Hi @cillian,
Thanks for this one! I am not sure I am trying to test the right thing here.

## What I did
1. Using this psql query I tried to find an order cycle with a lot of exchanges: `SELECT order_cycle_id, COUNT (order_cycle_id) FROM exchanges GROUP BY order_cycle_id ORDER BY count DESC;`
2. From the results I identified order cycle ID 3327 with 140 exchanges. 
![image](https://user-images.githubusercontent.com/1790176/203654202-e6527946-39ff-4df5-a0e2-6edada783ff6.png)
3. I opened `https://staging.openfoodnetwork.org.uk/admin/order_cycles/3327/edit` and `.../3327/incoming`, took screenshots and had a look at the number of requests and the loading time.

## Results
### Before
/edit: 12 requests, 53 seconds
/incoming: 150 requests, 57 seconds
### After
/edit: 12 requests, 56 seconds
/incoming: 151 requests, 59 seconds

- No reduction of loading time observed.
- No reduction of queries observed.
- The exchange queries are only sent on the /incoming page.

## Screenshots
### Before
/edit: 12 requests, 53 seconds
![image](https://user-images.githubusercontent.com/1790176/203655029-2dc1dee3-5c89-4565-8dac-04c738960d5b.png)

/incoming: 150 requests, 57 seconds
![image](https://user-images.githubusercontent.com/1790176/203655053-ebde1d43-27ef-4b3d-a0c0-873610d97e88.png)

### After
/edit: 12 requests, 56 seconds
![image](https://user-images.githubusercontent.com/1790176/203655162-3d9cd661-09fe-459a-ba32-a9a807db7f21.png)

/incoming: 151 requests, 59 seconds
![image](https://user-images.githubusercontent.com/1790176/203655194-138cd49f-7724-490f-8746-f807f0a68bb5.png)

## (Stupid) Questions
1. Am I looking at the right thing?
2. What exactly does an "exchange" mean?
3. When we talk about "tags" in this context, do we still mean the tags for showing/hiding order cycles, payment methods and so on to tagged customers?
4. How do "exchanges" and "tags" play together?

Thanks for providing some insight here!