@cillian 2022-11-25
Hi @drummer83 Thanks for testing this. It's the request to `/admin/order_cycles/3327.json` specifically that _should_ be improved by this PR. The exchanges are the incoming products coming in from the grower(s)/supplier(s) and the outgoing products from the distributor(s). Each _outgoing_ exchange can have tags like below...

![exchange-tags](https://user-images.githubusercontent.com/101088/204002588-ad942b0e-035e-4fdf-b404-1bd135dcb743.png)

The problem is that for _each and every_ outgoing exchange a database query is made to find which tags, if any, belong to that exchange. So if the order cycle has 7 outgoing exchanges for example (like in this [test](https://github.com/openfoodfoundation/openfoodnetwork/pull/9983/files#diff-4b05dd9f13b267021ea1adf1cc5f92150707a8bbb0573f18bdff24680d2af103R141)) there will be 7 database queries to lookup the tags. This PR changes things so it fetches the tags for all 7 outgoing exchanges in [a single query](https://github.com/openfoodfoundation/openfoodnetwork/pull/9983/files#diff-4b05dd9f13b267021ea1adf1cc5f92150707a8bbb0573f18bdff24680d2af103R147). I see the after time in your test was a bit slower at `54160ms` then the before time `52420ms`. I'm not sure why it's a bit slower yet, but it is good to investigate this further to check things are actually working faster, does that order cycle have any tags attached like in the screenshot above?

I tested it by loading the `/admin/order_cycles/3327.json` page directly in a browser rather than loading the `/edit` page and there seemed to be an improvement. I pulled these timings from the logs.

**Before:**

```
Completed 200 OK in 234ms (Views: 142.7ms | ActiveRecord: 81.2ms | Allocations: 93162)
Completed 200 OK in 243ms (Views: 150.7ms | ActiveRecord: 83.1ms | Allocations: 93161)
Completed 200 OK in 259ms (Views: 152.7ms | ActiveRecord: 92.1ms | Allocations: 93161)
Completed 200 OK in 226ms (Views: 137.2ms | ActiveRecord: 78.0ms | Allocations: 93161)
Completed 200 OK in 254ms (Views: 150.4ms | ActiveRecord: 88.8ms | Allocations: 93162)
Completed 200 OK in 290ms (Views: 173.5ms | ActiveRecord: 96.9ms | Allocations: 93164)
Completed 200 OK in 241ms (Views: 142.7ms | ActiveRecord: 87.0ms | Allocations: 93164)
Completed 200 OK in 234ms (Views: 141.9ms | ActiveRecord: 80.6ms | Allocations: 93161)
Completed 200 OK in 250ms (Views: 155.1ms | ActiveRecord: 84.8ms | Allocations: 93161)
Completed 200 OK in 266ms (Views: 160.9ms | ActiveRecord: 90.3ms | Allocations: 93161)
Average: 249.7ms
```

**After:**

```
Completed 200 OK in 206ms (Views: 128.2ms | ActiveRecord: 69.3ms | Allocations: 86780)
Completed 200 OK in 219ms (Views: 132.3ms | ActiveRecord: 75.8ms | Allocations: 86780)
Completed 200 OK in 213ms (Views: 131.5ms | ActiveRecord: 71.2ms | Allocations: 86782)
Completed 200 OK in 222ms (Views: 137.0ms | ActiveRecord: 74.8ms | Allocations: 86785)
Completed 200 OK in 216ms (Views: 135.2ms | ActiveRecord: 70.7ms | Allocations: 86780)
Completed 200 OK in 227ms (Views: 144.2ms | ActiveRecord: 75.7ms | Allocations: 86781)
Completed 200 OK in 211ms (Views: 129.2ms | ActiveRecord: 72.5ms | Allocations: 86780)
Completed 200 OK in 208ms (Views: 128.0ms | ActiveRecord: 71.7ms | Allocations: 86780)
Completed 200 OK in 261ms (Views: 153.1ms | ActiveRecord: 92.6ms | Allocations: 86780)
Completed 200 OK in 216ms (Views: 133.8ms | ActiveRecord: 73.5ms | Allocations: 86783)
Average: 219.9ms
```