@drummer83 2022-11-29
Hi @cillian,
Thanks for your explanations. I didn't use tags during my last test.

The strange thing is: On current master the loading time has suddenly dropped to only ~10 seconds (compared to > 50 seconds during my last test session). This is the same with and without a tag applied on the outgoing page and also after staging your PR. I am using the same order cycle as last time.
![image](https://user-images.githubusercontent.com/1790176/204671134-dddb4b5e-3f84-4693-9a80-601f2924772c.png)

Are we expecting to reduce the number of requests (13 in the example above - see bottom left corner) or are we expecting the request to be faster because within the request we have fewer database queries?

I have learned now that an exchange correlates to an enterprise and not to a product. :+1: 

Now I need to go to bed. I will try again. If you have any ideas or comments in meantime, it's much appreciated.
Thanks! And I'm sorry that I have difficulties testing this...