@drummer83 2022-12-07
Hi @filipefurtad0,

Sorry, I am totally stuck on this one. I still don't really understand the concept of exchanges and I don't know where to see the number of queries.

Could you maybe have a go at this one and document - as you always do - so I can learn for the next time?

I think it would be great to have it in this week's release, so I officially resign from this task. 😉