@filipefurtad0 2022-12-07
Hey @cillian @drummer83 ,

Thanks for all the testing and discussion. It's indeed not always easy to catch performance improvements in staging.

I understand an exchange each of the 4 lines (incoming) below:

![image](https://user-images.githubusercontent.com/49817236/206248835-a69360ee-9f61-4dd5-9537-e610d0d36d66.png)

For the outcoming section we have 3 exchanges:

![image](https://user-images.githubusercontent.com/49817236/206248972-5f0e463c-6b9f-44c7-8b13-3da7d9a0db0d.png)

So if we check the DB for this OC we have 7 exchanges (with tags):

```

openfoodnetwork=> select count(*) from exchanges where order_cycle_id = 3653;
 count 
-------
     7
(1 row)
```

Following the suggestion from @cillian (thanks!!), tracking the log of the staging server with:

`openfoodnetwork@staging:~/apps/openfoodnetwork/current$ tail -f log/staging.log | grep 'Completed 200 OK'`

And re-loading the page 10x we get:

**before this PR**

```

I, [2022-12-07 17:19:12#17778]  INFO -- : Completed 200 OK in 375ms (Views: 252.3ms | ActiveRecord: 107.3ms | Allocations: 94754)
I, [2022-12-07 17:19:14#17778]  INFO -- : Completed 200 OK in 363ms (Views: 244.6ms | ActiveRecord: 100.1ms | Allocations: 94756)
I, [2022-12-07 17:19:16#17472]  INFO -- : Completed 200 OK in 367ms (Views: 235.3ms | ActiveRecord: 117.0ms | Allocations: 94752)
I, [2022-12-07 17:19:17#17472]  INFO -- : Completed 200 OK in 361ms (Views: 238.4ms | ActiveRecord: 107.9ms | Allocations: 94758)
I, [2022-12-07 17:19:19#17472]  INFO -- : Completed 200 OK in 386ms (Views: 253.2ms | ActiveRecord: 117.4ms | Allocations: 94758)
I, [2022-12-07 17:19:21#17778]  INFO -- : Completed 200 OK in 416ms (Views: 273.5ms | ActiveRecord: 123.9ms | Allocations: 94752)
I, [2022-12-07 17:19:23#17778]  INFO -- : Completed 200 OK in 380ms (Views: 240.1ms | ActiveRecord: 118.1ms | Allocations: 94776)
I, [2022-12-07 17:19:25#17778]  INFO -- : Completed 200 OK in 396ms (Views: 256.8ms | ActiveRecord: 120.9ms | Allocations: 94753)
I, [2022-12-07 17:19:27#17778]  INFO -- : Completed 200 OK in 404ms (Views: 264.3ms | ActiveRecord: 122.2ms | Allocations: 94760)
I, [2022-12-07 17:19:29#17778]  INFO -- : Completed 200 OK in 430ms (Views: 281.6ms | ActiveRecord: 131.3ms | Allocations: 95575)
```

Average: 387.8 ms

**after this PR**

```
I, [2022-12-07 17:03:36#12537]  INFO -- : Completed 200 OK in 346ms (Views: 227.4ms | ActiveRecord: 102.4ms | Allocations: 90118)
I, [2022-12-07 17:03:38#12262]  INFO -- : Completed 200 OK in 339ms (Views: 222.4ms | ActiveRecord: 100.6ms | Allocations: 90117)
I, [2022-12-07 17:06:12#12262]  INFO -- : Completed 200 OK in 341ms (Views: 231.5ms | ActiveRecord: 93.6ms | Allocations: 90125)
I, [2022-12-07 17:06:15#12537]  INFO -- : Completed 200 OK in 327ms (Views: 213.0ms | ActiveRecord: 96.0ms | Allocations: 90128)
I, [2022-12-07 17:06:17#12262]  INFO -- : Completed 200 OK in 359ms (Views: 230.4ms | ActiveRecord: 111.1ms | Allocations: 90507)
I, [2022-12-07 17:06:19#12262]  INFO -- : Completed 200 OK in 340ms (Views: 222.6ms | ActiveRecord: 99.1ms | Allocations: 90117)
I, [2022-12-07 17:06:21#12537]  INFO -- : Completed 200 OK in 309ms (Views: 196.7ms | ActiveRecord: 96.5ms | Allocations: 90117)
I, [2022-12-07 17:06:23#12537]  INFO -- : Completed 200 OK in 350ms (Views: 239.4ms | ActiveRecord: 95.4ms | Allocations: 90124)
I, [2022-12-07 17:06:26#12537]  INFO -- : Completed 200 OK in 353ms (Views: 232.8ms | ActiveRecord: 104.1ms | Allocations: 90118)
I, [2022-12-07 17:06:30#12537]  INFO -- : Completed 200 OK in 343ms (Views: 229.1ms | ActiveRecord: 98.2ms | Allocations: 90123)
```

Average: 340.7 ms

So the improvement is indeed noticeable and reproducible in staging :+1: 

Also tested basic tag functionality: hiding an OC by default - OK!

Merging :tada: